src_fr/requirements.txt
# Bibliothèques requises pour le système de détection AML
pandas>=2.0.0
numpy>=2.0.0
pyyaml>=6.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
tqdm>=4.65.0 